Data Logging barebone test
Ironed out data logger and tested with a mock arduino on the beaglebone
Logging works as expected, however may need to change time stamp to
record the amount of time since the program started.  This may need to
take into account the time at which the pig begins to move.

Also added some comments to code to help explain what is happening.

For now the beaglebone setup has been commented out for testing
purposes.