JNA Wrapper
The JNA wrapper works.  Curently able to send a single float through the
I2C lines which is taking roughly 1 millisecond or less per transfer.
There is no read functionality provided here.

There is a native c file that this uses currently called tmp102.so.
This file has to be compiled on the beaglebone.  The command to do so
is: gcc -fPIC tmp102.c -o tmp102.so -shared